Bug fix!, run small tasks on small cores and large tasks on large cores.
Don't do work stealing, if we run large functions on small cores then there
is a chance of head of line blocking. And if we run small functions on
large cores then there is chance that completion times for such tasks will
be undefined due to out of sync timestamp counter.